OrchidOs Security : Dev,  Old target  and Adb notifications
Informing the user that ADB is enabled does not necessitate an alert
sound. It's especially annoying with the default AOSP notification sound,
but even with a less distracting sound, this doesn't need to be so
intrusive.

Change-Id: Idee92b44cff2473a3686a2f2d95c6f5e4e91c2f8

Allow disabling important developer notifications

It looks like this notification channel was *supposed* to be blockable,
but the setBlockable call wasn't updated after copying the developer
channel.

Either way, let's allow the user to disable these notifications to reduce
clutter in the notification shade.

Change-Id: I4eef81537af74f451f7d2163223346876d68ecc4
Old target
core: Remove old app target SDK dialog

If an app is old, users should already know that, and there's usually no
point in warning them about it because they would already be using a
newer version if one existed. Sometimes, using an old app is necessary
for one reason or another, so remove this annoyance and let the user use
their old app in peace.

Change-Id: I1a3021f0f9bec1ab6ff7641776391e1bd4c3cc49