OrchidOs UI :  Packages :  SystemUI : Remove build number
SystemUI: Remove Android build number from QS footer

This is only shown when developer settings is enabled, but we developers
tend to always have it enabled. Irrelevant text showing up in the QS
footer is a bit ugly and it's not really useful when this information is
easily accessible in Settings anyway.

Change-Id: Ideb38eeb603dc5c51945fafeae9c4eca18531b5d